Ready permissions

+++ src/main/java/dev/ivantolkach/kanban/KanbanBoardServer/application/service/TagService.java
@@ -5,10 +5,9 @@ import dev.ivantolkach.kanban.KanbanBoardServer.application.dto.tag.TagDTOOutput
 import dev.ivantolkach.kanban.KanbanBoardServer.application.dto.tag.TagFilterDTO;
 import dev.ivantolkach.kanban.KanbanBoardServer.application.dto.tasktag.TaskTagDTO;
 import dev.ivantolkach.kanban.KanbanBoardServer.application.service.exception.NotFoundException;
-import dev.ivantolkach.kanban.KanbanBoardServer.domain.model.Project;
-import dev.ivantolkach.kanban.KanbanBoardServer.domain.model.Tag;
-import dev.ivantolkach.kanban.KanbanBoardServer.domain.model.Task;
-import dev.ivantolkach.kanban.KanbanBoardServer.domain.model.TaskTag;
+import dev.ivantolkach.kanban.KanbanBoardServer.application.service.exception.UnauthorizedException;
+import dev.ivantolkach.kanban.KanbanBoardServer.domain.common.enums.UserRole;
+import dev.ivantolkach.kanban.KanbanBoardServer.domain.model.*;
 import dev.ivantolkach.kanban.KanbanBoardServer.infrastructure.mapper.tag.TagListMapper;
 import dev.ivantolkach.kanban.KanbanBoardServer.infrastructure.mapper.tag.TagMapper;
 import dev.ivantolkach.kanban.KanbanBoardServer.infrastructure.mapper.tasktag.TaskTagListMapper;
@@ -19,6 +18,8 @@ import dev.ivantolkach.kanban.KanbanBoardServer.infrastructure.persistence.repos
 import dev.ivantolkach.kanban.KanbanBoardServer.infrastructure.persistence.specification.TagSpecification;
 import dev.ivantolkach.kanban.KanbanBoardServer.infrastructure.persistence.specification.TaskTagSpecification;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -48,6 +49,9 @@ public class TagService {
     @Autowired
     TaskRepository taskRepository;
 
+    @Autowired
+    UserService userService;
+
     public boolean existsById(UUID tagId) {
         return tagRepository.existsById(tagId);
     }
@@ -57,16 +61,34 @@ public class TagService {
     }
 
     public List<TaskTagDTO> getTaskTagsByFilter(TaskTagDTO filter) {
-        return taskTagListMapper.toDTOList(taskTagRepository.findAll(TaskTagSpecification.filterBy(filter)));
+        User currentUser = userService.getCurrentUser();
+
+        Specification<TaskTag> spec = TaskTagSpecification.filterBy(filter);
+
+        if (currentUser.getRole() != UserRole.ROLE_ADMIN) {
+            spec = spec.and(TaskTagSpecification.accessibleBy(currentUser));
+        }
+
+        List<TaskTag> taskTags = taskTagRepository.findAll(spec);
+        return taskTagListMapper.toDTOList(taskTags);
     }
 
     public TagDTOOutput createUpdateTag(TagDTOInput tagDTOInput) {
+        User currentUser = userService.getCurrentUser();
 
         Tag tag;
 
         if (tagDTOInput.getId() != null) {
             tag = tagRepository.findById(tagDTOInput.getId())
                     .orElseThrow(()-> new NotFoundException("Tag not found with id: " + tagDTOInput.getId()));
+
+            if (!(tagDTOInput.getName() == null || tagDTOInput.getName().isBlank())) {
+                tag.setName(tagDTOInput.getName());
+            }
+
+            if (currentUser.getRole() != UserRole.ROLE_ADMIN && !tag.getCreatedBy().getId().equals(currentUser.getId())) {
+                throw new UnauthorizedException("Not allowed to edit this entity");
+            }
 
             if (!(tagDTOInput.getName() == null || tagDTOInput.getName().isBlank())) {
                 tag.setName(tagDTOInput.getName());
@@ -80,11 +102,43 @@ public class TagService {
     }
 
     public TaskTagDTO attachTag (UUID taskId, UUID tagId) {
+        User currentUser = userService.getCurrentUser();
+        if (currentUser == null) {
+            throw new AccessDeniedException("User is not authenticated");
+        }
+
         Task task = taskRepository.findById(taskId)
                 .orElseThrow(()-> new NotFoundException("Task not found with id: " + taskId));
 
         Tag tag = tagRepository.findById(tagId)
                 .orElseThrow(()-> new NotFoundException("Tag not found with id: " + tagId));
+
+        if (taskTagRepository.existsByTaskIdAndTagId(taskId, tagId)) {
+            throw new IllegalStateException("Tag " + tagId + " is already attached to task " + taskId);
+        }
+
+        if (currentUser.getRole() != UserRole.ROLE_ADMIN) {
+
+            boolean isTaskAuthor =
+                    task.getCreatedBy() != null &&
+                            task.getCreatedBy().getId().equals(currentUser.getId());
+
+            ProjectColumn column = task.getColumn();
+            boolean isColumnAuthor =
+                    column.getCreatedBy() != null &&
+                            column.getCreatedBy().getId().equals(currentUser.getId());
+
+            Project project = column.getProject();
+            boolean isProjectAuthor =
+                    project.getCreatedBy() != null &&
+                            project.getCreatedBy().getId().equals(currentUser.getId());
+
+            if (!isTaskAuthor && !isColumnAuthor && !isProjectAuthor) {
+                throw new AccessDeniedException(
+                        "You do not have permission to attach a tag to this task"
+                );
+            }
+        }
 
         if (taskTagRepository.existsByTaskIdAndTagId(taskId, tagId)) {
             throw new IllegalStateException("Tag " + tagId + " is already attached to task " + taskId);
@@ -98,15 +152,29 @@ public class TagService {
     }
 
     public void deleteTag(UUID tagId) {
+        User currentUser = userService.getCurrentUser();
+
         Tag existingTag = tagRepository.findById(tagId)
-                .orElseThrow(()-> new NotFoundException("Tag not found with id: " + tagId));
+                .orElseThrow(() -> new NotFoundException("Tag not found with id: " + tagId));
+
+        if (currentUser.getRole() != UserRole.ROLE_ADMIN && !existingTag.getCreatedBy().getId().equals(currentUser.getId())) {
+            throw new UnauthorizedException("Not allowed to delete this entity");
+        }
 
         tagRepository.delete(existingTag);
     }
 
     public void deleteTaskTag(UUID taskTagId) {
+        User currentUser = userService.getCurrentUser();
+
         TaskTag existingTaskTag = taskTagRepository.findById(taskTagId)
-                .orElseThrow(()->new NotFoundException("TaskTag relation not found with id: " + taskTagId));
+                .orElseThrow(() -> new NotFoundException("TaskTag relation not found with id: " + taskTagId));
+
+        Task task = existingTaskTag.getTask();
+
+        if (currentUser.getRole() != UserRole.ROLE_ADMIN && !task.getCreatedBy().getId().equals(currentUser.getId())) {
+            throw new UnauthorizedException("Not allowed to delete this relation");
+        }
 
         taskTagRepository.delete(existingTaskTag);
     }
